Rename getKeyphraseLength to getWordcount
