+++ js/analyses/getWordCount.js
@@ -3,10 +3,10 @@
 var sanitizeString = require( "../stringProcessing/sanitizeString.js" );
 
 /**
- * Checks the number of words in the keyphrase
+ * Checks the number of words in a string
  *
  * @param {string} text The keyphrase to count words in.
- * @returns {number} The wordcount of the given keyphrase.
+ * @returns {number} The wordcount of the given string.
  */
 module.exports = function( text ) {
 	text = sanitizeString( text );
